// types/constructor.ts
export interface Constructor {
  position: number;
  points: number;
  wins: number;
  id: string;
  name: string;
  color: string;
  logo: string;
  fullName: string;
  base: string;
  entryYear: number;
  championships: number;
  seasonWins: number;
  chassis: string;
  powerUnit: string;
  teamChief: string;
  technicalChief: string;
}